# perf(rust, python): speed up csv parsing for slower datetimes formats