--- a/polars/polars-io/src/csv/buffer.rs
+++ b/polars/polars-io/src/csv/buffer.rs
@@ -423,15 +423,11 @@ where
         return Ok(());
     };
 
-    match infer_pattern_single(val) {
-        None => {
-            buf.builder.append_null();
-            Ok(())
-        }
-        Some(pattern_with_offset) => {
-            match DatetimeInfer::<T::Native>::try_from(pattern_with_offset.pattern) {
+    match &buf.compiled {
+        Some(compiled) => {
+            match DatetimeInfer::<T::Native>::try_from(compiled.pattern_with_offset.pattern) {
                 Ok(mut infer) => {
-                    let parsed = infer.parse(val, pattern_with_offset.offset);
+                    let parsed = infer.parse(val, compiled.pattern_with_offset.offset);
                     buf.compiled = Some(infer);
                     buf.builder.append_option(parsed);
                     Ok(())
@@ -439,6 +435,28 @@ where
                 Err(_) => {
                     buf.builder.append_null();
                     Ok(())
+                }
+            }
+        }
+        None => {
+            match infer_pattern_single(val) {
+                None => {
+                    buf.builder.append_null();
+                    Ok(())
+                }
+                Some(pattern_with_offset) => {
+                    match DatetimeInfer::<T::Native>::try_from(pattern_with_offset.pattern) {
+                        Ok(mut infer) => {
+                            let parsed = infer.parse(val, pattern_with_offset.offset);
+                            buf.compiled = Some(infer);
+                            buf.builder.append_option(parsed);
+                            Ok(())
+                        }
+                        Err(_) => {
+                            buf.builder.append_null();
+                            Ok(())
+                        }
+                    }
                 }
             }
         }