--- a/polars/polars-io/src/csv/utils.rs
+++ b/polars/polars-io/src/csv/utils.rs
@@ -123,10 +123,7 @@ fn infer_field_schema(string: &str, try_parse_dates: bool) -> DataType {
                         PatternWithOffset {
                             pattern: Pattern::DatetimeYMDZ,
                             offset,
-                        } => DataType::Datetime(
-                            TimeUnit::Microseconds,
-                            offset.map(|x| x.to_string()),
-                        ),
+                        } => DataType::Utf8,  // TODO: support this
                     },
                     None => DataType::Utf8,
                 }
@@ -162,7 +159,7 @@ fn infer_field_schema(string: &str, try_parse_dates: bool) -> DataType {
                     PatternWithOffset {
                         pattern: Pattern::DatetimeYMDZ,
                         offset,
-                    } => DataType::Datetime(TimeUnit::Microseconds, offset.map(|x| x.to_string())),
+                    } => DataType::Utf8,  // TODO: support this
                 },
                 None => DataType::Utf8,
             }