--- a/polars/polars-time/src/chunkedarray/utf8/infer.rs
+++ b/polars/polars-time/src/chunkedarray/utf8/infer.rs
@@ -80,6 +80,7 @@ impl StrpTimeParser<i32> for DatetimeInfer<i32> {
 
 #[derive(Clone)]
 pub struct DatetimeInfer<T> {
+    pub pattern_with_offset: PatternWithOffset,
     patterns: &'static [&'static str],
     latest_fmt: &'static str,
     transform: fn(&str, &str, Option<FixedOffset>, bool) -> Option<T>,
@@ -96,6 +97,7 @@ impl TryFrom<Pattern> for DatetimeInfer<i64> {
     fn try_from(value: Pattern) -> PolarsResult<Self> {
         match value {
             Pattern::DatetimeDMY => Ok(DatetimeInfer {
+                pattern_with_offset: PatternWithOffset{pattern: Pattern::DatetimeDMY, offset: None},
                 patterns: patterns::DATETIME_D_M_Y,
                 latest_fmt: patterns::DATETIME_D_M_Y[0],
                 transform: transform_datetime_us,
@@ -105,6 +107,7 @@ impl TryFrom<Pattern> for DatetimeInfer<i64> {
                 utc: false,
             }),
             Pattern::DatetimeYMD => Ok(DatetimeInfer {
+                pattern_with_offset: PatternWithOffset{pattern:Pattern::DatetimeYMD, offset:None},
                 patterns: patterns::DATETIME_Y_M_D,
                 latest_fmt: patterns::DATETIME_Y_M_D[0],
                 transform: transform_datetime_us,
@@ -114,6 +117,7 @@ impl TryFrom<Pattern> for DatetimeInfer<i64> {
                 utc: false,
             }),
             Pattern::DatetimeYMDZ => Ok(DatetimeInfer {
+                pattern_with_offset: PatternWithOffset{pattern:Pattern::DatetimeYMDZ, offset:None},
                 patterns: patterns::DATETIME_Y_M_D_Z,
                 latest_fmt: patterns::DATETIME_Y_M_D_Z[0],
                 transform: transform_tzaware_datetime_us,
@@ -134,6 +138,7 @@ impl TryFrom<Pattern> for DatetimeInfer<i32> {
     fn try_from(value: Pattern) -> PolarsResult<Self> {
         match value {
             Pattern::DateDMY => Ok(DatetimeInfer {
+                pattern_with_offset: PatternWithOffset{pattern: Pattern::DateDMY, offset: None},
                 patterns: patterns::DATE_D_M_Y,
                 latest_fmt: patterns::DATE_D_M_Y[0],
                 transform: transform_date,
@@ -143,6 +148,7 @@ impl TryFrom<Pattern> for DatetimeInfer<i32> {
                 utc: false,
             }),
             Pattern::DateYMD => Ok(DatetimeInfer {
+                pattern_with_offset: PatternWithOffset{pattern: Pattern::DateYMD, offset: None},
                 patterns: patterns::DATE_Y_M_D,
                 latest_fmt: patterns::DATE_Y_M_D[0],
                 transform: transform_date,