For the following:
```python
from datetime import datetime, date
import datetime as dt

import polars as pl
import io

dates = pl.date_range(datetime(1000, 1, 1), datetime(9999, 1, 1)).dt.strftime('%m-%d-%Y %H:%M:%S')
data = 'my_column\\n' + '\\n'.join(dates)
print(ipython.run_line_magic("timeit", "pl.read_csv(io.StringIO(data), rechunk=False, try_parse_dates=True)"))

dates = pl.date_range(date(1000, 1, 1), date(9999, 1, 1)).dt.strftime('%m-%d-%Y')
data = 'my_column\\n' + '\\n'.join(dates)
print(ipython.run_line_magic("timeit", "pl.read_csv(io.StringIO(data), rechunk=False, try_parse_dates=True)"))

# current happy-paths
dates = pl.date_range(datetime(1000, 1, 1), datetime(9999, 1, 1)).dt.strftime('%d-%m-%Y %H:%M:%S')
data = 'my_column\\n' + '\\n'.join(dates)
print(ipython.run_line_magic("timeit", "pl.read_csv(io.StringIO(data), rechunk=False, try_parse_dates=True)"))
```
here are some timings https://www.kaggle.com/code/marcogorelli/polars-timing?scriptVersionId=125707108

this branch:
- 5.86 s ± 170 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 760 ms ± 3.12 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 540 ms ± 4.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

upstream/main:
- 27.2 s ± 140 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 25.7 s ± 191 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 593 ms ± 47.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

If the fail-fast conditions can be sorted out too, then it should become even better. But this already solves a large part of the issue