@ritchie46 2023-04-14
Could you explain a bit how the terrible performance was fixed for my mental model?