@MarcoGorelli 2023-04-14
Sure - the issue is that before, if `parse_bytes` failed, then `infer_pattern_single` would be called within `slow_datetime_parser` for _each_ element

With this PR, the pattern is passed down, and then only formats from that pattern can be tried. If this were combined with fail-fast conditions from #8111, then it would get ever better

It is an issue to make `pattern_with_offset` a public attribute of `DatetimeInfer`? If so, I'll think if there's another way to pass the pattern to `slow_datetime_parser`