@ritchie46 2023-04-14
> With this PR, the pattern is passed down, and then only formats from that pattern can be tried. If this were combined with fail-fast conditions from https://github.com/pola-rs/polars/pull/8111, then it would get ever better

Ah, right, that makes sense. :+1:  Thanks!

> It is an issue to make pattern_with_offset a public attribute of DatetimeInfer? If so, I'll think if there's another way to pass the pattern to slow_datetime_parser

No that's fine. Another solution often is exposing a function that returns a reference to that element. But there is not unsafe involved here, so I think this is fine.

Thanks a lot for this one, this was horrific performance.